+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fixed issue with textfield tooltip not appearing
 - Possible fix for configs restore issue
 - Implemented new option "Invert color on dark page" + custom background color on dark pages
+- Improvements for textfield handling on vk.com and yandex.ru
 - Other small fixes and improvements
 
 #3.3.2